refactor(viewer): modify import location to mock data
