+++ libs/fixtures/src/annotation/contours.ts
@@ -1,4 +1,4 @@
-import { PolygonAnnotation } from '@lunit/insight-viewer'
+import { PolygonAnnotation } from '@lunit/insight-viewer/annotation'
 
 export const CONTOURS: PolygonAnnotation[] = [
   {
