+++ libs/insight-viewer/src/Annotation/index.ts
@@ -1,3 +1,4 @@
 export { AnnotationOverlay } from './components/AnnotationOverlay'
+export * from './types'
 
 export type { Annotation, DrawableAnnotation, AnnotationMode } from '../Annotation/types'
