--- libs/insight-viewer/src/mocks/circle.ts
@@ -1,11 +1,0 @@
-import { CircleAnnotation } from '../types'
-
-const INITIAL_CIRCLE: CircleAnnotation = {
-  id: 1,
-  lineWidth: 1.5,
-  type: 'circle',
-  center: [319.6342857142857, 186.7085714285714],
-  radius: 0,
-}
-
-export { INITIAL_CIRCLE }
