+++ libs/insight-viewer/src/mocks/utils.ts
@@ -1,3 +1,3 @@
-import { Point } from '../types'
+import { Point } from '../Annotation/types'
 
 export const mockPixelToCanvas = ([x, y]: Point): Point => [x, y]
